requirements.txt
gspread
oauth2client
requests
beautifulsoup4
openai
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib